+++ logistics/plugins/__init__.py
@@ -1,30 +1,6 @@
 # imports relevant contents.
 from logistics.plugins.metaclass import Meta
-from logistics.plugins.types import (
-    VandalType,
-    IntegerType,
-    FloatType,
-    NumberType,
-    ReturnType,
-    PrintType,
-    GraphType,
-    StringType,
-    ListType,
-    TupleType,
-    DictionaryType,
-    BooleanType,
-    FilePathType,
-    SpecialType,
-    NumberVector,
-    StringVector,
-    StringDictionary,
-    DictionaryVector,
-    NumberVectorAlike,
-    NumberArrayAlike,
-    AnyArrayAlike,
-    AnyVectorAlike,
-    AnyType,
-)
+from logistics.plugins.types import *
 
 # all relevant contents.
 __all__ = [
@@ -51,34 +27,13 @@ __all__ = [
     'AnyArrayAlike',
     'AnyVectorAlike',
     'AnyType',
+    'VandalTypes',
     'Meta',
 ]
 
 # all available types.
-types = [
-    VandalType,
-    IntegerType,
-    FloatType,
-    NumberType,
-    ReturnType,
-    PrintType,
-    GraphType,
-    StringType,
-    ListType,
-    TupleType,
-    DictionaryType,
-    BooleanType,
-    FilePathType,
-    SpecialType,
-    NumberVector,
-    StringVector,
-    StringDictionary,
-    DictionaryVector,
-    NumberVectorAlike,
-    NumberArrayAlike,
-    AnyArrayAlike,
-    AnyVectorAlike,
-    AnyType,
+VandalTypes = [
+    'VandalTypes',
 ]
 
 # metaclass.